Benachrichtugungen/Fehlermeldungen anzeigen in eigenes Addon
für #1265
