+++ bundles/aero.minova.rcp.rcp/fragment.e4xmi
@@ -381,5 +381,6 @@
   </fragments>
   <fragments xsi:type="fragment:StringModelFragment" xmi:id="_G_OTECEvEeyLzPNotbg_ew" featurename="addons" parentElementId="xpath:/">
     <elements xsi:type="application:Addon" xmi:id="_Lx0fYCEvEeyLzPNotbg_ew" elementId="aero.minova.rcp.rcp.addon.iconrendering" contributionURI="bundleclass://aero.minova.rcp.rcp/aero.minova.rcp.rcp.addons.IconRendering"/>
+    <elements xsi:type="application:Addon" xmi:id="_LiQbIPYfEeyxFrZzmcAvXw" elementId="aero.minova.rcp.rcp.addon.notificationanderroraddon" contributionURI="bundleclass://aero.minova.rcp.rcp/aero.minova.rcp.rcp.addons.NotificationAndErrorAddon"/>
   </fragments>
 </fragment:ModelFragments>
