+++ bundles/aero.minova.rcp.rcp/src/aero/minova/rcp/rcp/util/WFCDetailCASRequestsUtil.java
@@ -1,11 +1,6 @@
 package aero.minova.rcp.rcp.util;
 
-import java.text.MessageFormat;
-import java.text.NumberFormat;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -30,15 +25,12 @@ import org.eclipse.e4.ui.di.UIEventTopic;
 import org.eclipse.e4.ui.di.UISynchronize;
 import org.eclipse.e4.ui.model.application.ui.advanced.MPerspective;
 import org.eclipse.e4.ui.model.application.ui.basic.MWindow;
-import org.eclipse.e4.ui.services.IServiceConstants;
 import org.eclipse.e4.ui.workbench.UIEvents;
 import org.eclipse.e4.ui.workbench.modeling.EModelService;
-import org.eclipse.e4.ui.workbench.modeling.EPartService;
 import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.swt.layout.FormLayout;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Display;
-import org.eclipse.swt.widgets.Shell;
 
 import aero.minova.rcp.constants.Constants;
 import aero.minova.rcp.css.widgets.MinovaSection;
@@ -75,11 +67,8 @@ import aero.minova.rcp.preferences.ApplicationPreferences;
 import aero.minova.rcp.rcp.accessor.AbstractValueAccessor;
 import aero.minova.rcp.rcp.accessor.GridAccessor;
 import aero.minova.rcp.rcp.accessor.SectionAccessor;
-import aero.minova.rcp.rcp.handlers.ShowErrorDialogHandler;
 import aero.minova.rcp.rcp.parts.WFCDetailPart;
 import aero.minova.rcp.rcp.widgets.SectionGrid;
-import aero.minova.rcp.util.DateUtil;
-import aero.minova.rcp.widgets.MinovaNotifier;
 
 public class WFCDetailCASRequestsUtil {
 
@@ -110,22 +99,7 @@ public class WFCDetailCASRequestsUtil {
 	EModelService model;
 
 	@Inject
-	EPartService partService;
-
-	@Inject
 	IEventBroker broker;
-
-	@Inject
-	@Named(IServiceConstants.ACTIVE_SHELL)
-	private Shell shell;
-
-	@Inject
-	@Preference(nodePath = "aero.minova.rcp.preferencewindow", value = "user")
-	String employee;
-
-	@Inject
-	@Preference(nodePath = "aero.minova.rcp.preferencewindow", value = "timezone")
-	String timezone;
 
 	@Inject
 	@Preference(nodePath = ApplicationPreferences.PREFERENCES_NODE, value = ApplicationPreferences.AUTO_RELOAD_INDEX)
@@ -543,7 +517,7 @@ public class WFCDetailCASRequestsUtil {
 				return null;
 			});
 		} else {
-			openNotificationPopup("msg.ActionAborted");
+			broker.send(Constants.BROKER_SHOWNOTIFICATION, "msg.ActionAborted");
 		}
 	}
 
@@ -567,10 +541,10 @@ public class WFCDetailCASRequestsUtil {
 		SqlProcedureResult mainResult = resultList.get(0).getSQLProcedureResult();
 		if (insert) {
 			setKeysFromTable(mainResult.getOutputParameters());
-			openNotificationPopup("msg.DataSaved");
+			broker.send(Constants.BROKER_SHOWNOTIFICATION, "msg.DataSaved");
 			handleUserAction(Constants.INSERT_REQUEST, mainResult.getOutputParameters());
 		} else {
-			openNotificationPopup("msg.DataUpdated");
+			broker.send(Constants.BROKER_SHOWNOTIFICATION, "msg.DataUpdated");
 			handleUserAction(Constants.UPDATE_REQUEST, currentKeyTable);
 		}
 
@@ -611,146 +585,6 @@ public class WFCDetailCASRequestsUtil {
 
 		sendEventToHelper(ActionCode.AFTERSAVE);
 		focusFirstEmptyField();
-	}
-
-	/**
-	 * Liefert das übersetzte Objekt zurück
-	 *
-	 * @param translate
-	 * @return
-	 */
-	private String getTranslation(String translate) {
-		if (!translate.startsWith("@")) {
-			translate = "@" + translate;
-		}
-		return translationService.translate(translate, null);
-	}
-
-	@Inject
-	@Optional
-	public void showErrorMessage(@UIEventTopic(Constants.BROKER_SHOWERRORMESSAGE) String message) {
-		MPerspective activePerspective = model.getActivePerspective(partContext.get(MWindow.class));
-		if (activePerspective.equals(perspective)) {
-			// Fokus auf den Search Part legen, damit Fehlermeldungen nicht mehrmals angezeigt werden
-			String commandID = Constants.AERO_MINOVA_RCP_RCP_COMMAND_SELECTSEARCHPART;
-			ParameterizedCommand cmd = commandService.createCommand(commandID, null);
-			handlerService.executeHandler(cmd);
-
-			MessageDialog.openError(shell, getTranslation(ERROR), getTranslation(message));
-		}
-	}
-
-	@Inject
-	@Optional
-	public void showErrorMessage(@UIEventTopic(Constants.BROKER_SHOWERROR) ErrorObject et) {
-		MPerspective activePerspective = model.getActivePerspective(partContext.get(MWindow.class));
-		if (activePerspective.equals(perspective)) {
-
-			String value = formatMessage(et);
-			value += "\n\nUser : " + et.getUser();
-			value += "\nProcedure/View: " + et.getProcedureOrView();
-
-			// Fokus auf den Search Part legen, damit Fehlermeldungen von Lookups nicht mehrmals angezeigt werden
-			String commandID = Constants.AERO_MINOVA_RCP_RCP_COMMAND_SELECTSEARCHPART;
-			ParameterizedCommand cmd = commandService.createCommand(commandID, null);
-			handlerService.executeHandler(cmd);
-
-			if (et.getT() == null) {
-				MessageDialog.openError(shell, getTranslation(ERROR), value);
-			} else {
-				ShowErrorDialogHandler.execute(shell, getTranslation(ERROR), value, et.getT());
-			}
-		}
-	}
-
-	private String formatMessage(ErrorObject et) {
-		Table errorTable = et.getErrorTable();
-
-		if (errorTable == null || errorTable.getRows().get(0).getValue(0) == null) {
-			return "Internal Server Error";
-		}
-
-		Value vMessageProperty = errorTable.getRows().get(0).getValue(0);
-		String messageproperty = vMessageProperty.getStringValue().strip();
-		String value = translationService.translate("@" + messageproperty, null);
-
-		if (value.equals(messageproperty) && errorTable.getColumnIndex("DEFAULT") != -1) { // Keine Übersetzung gefunden -> Default nutzen
-			value = errorTable.getRows().get(0).getValue(errorTable.getColumnIndex("DEFAULT")).getStringValue();
-		} else if (errorTable.getColumnCount() > 1) {// Ticket number {0} is not numeric
-
-			value = value.replaceAll("%(\\d*)", "{$1}"); // %n fürs Formattieren mit {n} ersetzen
-
-			List<String> params = new ArrayList<>();
-			for (int i = 1; i < errorTable.getColumnCount(); i++) {
-				Value v = errorTable.getRows().get(0).getValue(i);
-				String columnName = errorTable.getColumnName(i);
-				switch (columnName) {
-				case "p":
-					params.add(translationService.translate("@" + v.getStringValue(), null));
-					break;
-				case "i":
-					try {
-						params.add("" + NumberFormat.getInstance(wfcDetailPart.getLocale()).parse(v.getStringValue()).intValue());
-					} catch (ParseException e) {
-						e.printStackTrace();
-					}
-					break;
-				case "f.iso":
-					params.add("" + Float.parseFloat(v.getStringValue()));
-					break;
-				case "f":
-					try {
-						params.add("" + NumberFormat.getInstance(wfcDetailPart.getLocale()).parse(v.getStringValue()).floatValue());
-					} catch (ParseException e) {
-						e.printStackTrace();
-					}
-					break;
-				case "d.iso":
-					try {
-						Date parsedDate = new SimpleDateFormat("yyyyMMdd").parse(v.getStringValue());
-						params.add(DateUtil.getDateString(parsedDate.toInstant(), wfcDetailPart.getLocale(), null));
-					} catch (ParseException e) {
-						e.printStackTrace();
-					}
-					break;
-				case "d":
-					try {
-						Date parsedDate = new SimpleDateFormat("ddMMyyyy").parse(v.getStringValue());
-						params.add(DateUtil.getDateString(parsedDate.toInstant(), wfcDetailPart.getLocale(), null));
-					} catch (ParseException e) {
-						e.printStackTrace();
-					}
-					break;
-				case "DEFAULT": // Spalte mit DEFAULT-String -> kein Param
-					break;
-				default:
-					// "s" oder ungültiger Spaltenname -> String
-					params.add(v.getStringValue());
-					break;
-				}
-			}
-			value = MessageFormat.format(value, params.toArray(new String[0]));
-		}
-
-		return value;
-	}
-
-	@Inject
-	@Optional
-	public void showNotification(@UIEventTopic(Constants.BROKER_SHOWNOTIFICATION) String message) {
-		MPerspective activePerspective = model.getActivePerspective(partContext.get(MWindow.class));
-		if (activePerspective.equals(perspective)) {
-			openNotificationPopup(message);
-		}
-	}
-
-	@Inject
-	@Optional
-	public void showNotification(@UIEventTopic(Constants.BROKER_SHOWNOTIFICATION) ErrorObject et) {
-		MPerspective activePerspective = model.getActivePerspective(partContext.get(MWindow.class));
-		if (activePerspective.equals(perspective)) {
-			openNotificationPopup(formatMessage(et));
-		}
 	}
 
 	@Inject
@@ -854,7 +688,7 @@ public class WFCDetailCASRequestsUtil {
 			ParameterizedCommand cmd = commandService.createCommand(Constants.AERO_MINOVA_RCP_RCP_COMMAND_LOADINDEX, null);
 			handlerService.executeHandler(cmd);
 		}
-		openNotificationPopup("msg.DataDeleted");
+		broker.send(Constants.BROKER_SHOWNOTIFICATION, "msg.DataDeleted");
 		Map<MPerspective, String> map = new HashMap<>();
 		map.put(perspective, Constants.DELETE_REQUEST);
 		clearFields(map);
@@ -885,8 +719,8 @@ public class WFCDetailCASRequestsUtil {
 			tableFuture.thenAccept(ta -> sync.syncExec(() -> {
 				broker.send(Constants.BROKER_PROCEDUREWITHTABLESUCCESS, ta);
 				if (ta != null && ta.getResultSet() != null && ERROR.equals(ta.getResultSet().getName())) {
-					ErrorObject e = new ErrorObject(ta.getResultSet(), "USER");
-					showErrorMessage(e);
+					ErrorObject e = new ErrorObject(ta.getResultSet(), dataService.getUserName());
+					broker.send(Constants.BROKER_SHOWERROR, e);
 				} else if (ta != null) {
 					selectedTable = ta.getResultSet();
 					if (!selectedTable.getRows().isEmpty()) {
@@ -897,17 +731,6 @@ public class WFCDetailCASRequestsUtil {
 				}
 				broker.send(Constants.BROKER_PROCEDUREWITHTABLESUCCESSFINISHED, ta);
 			}));
-		}
-	}
-
-	/**
-	 * Öffet ein Popup, welches dem Nutzer über den Erfolg oder das Scheitern seiner Anfrage informiert
-	 *
-	 * @param message
-	 */
-	public void openNotificationPopup(String message) {
-		if (!shell.getDisplay().isDisposed()) {
-			MinovaNotifier.show(shell, getTranslation(message), getTranslation("Notification"));
 		}
 	}
